Address some incompatibilities between Scala 2 and 3

+++ src/test/yuck/core/test/DomainTestHelper.scala
@@ -29,8 +29,8 @@ class DomainTestHelper[Value <: AnyValue](logger: LazyLogger) extends YuckAssert
         def checkDistribution(f: Map[Value, Int]): Unit = {
             for (a <- d.values) {
                 import scala.math.Ordering.Double.TotalOrdering
-                assertGt(f.getOrElse(a, 0), sampleSize / d.size * (1 - maxError))
-                assertLt(f.getOrElse(a, 0), sampleSize / d.size * (1 + maxError))
+                assertGt(f.getOrElse(a, 0).toDouble, sampleSize / d.size * (1 - maxError))
+                assertLt(f.getOrElse(a, 0).toDouble, sampleSize / d.size * (1 + maxError))
             }
         }
         val f1 = new mutable.AnyRefMap[Value, Int]
